requirements.txt
python-binance
pandas
numpy
mplfinance